fix trivial compiler warning for last commit

+++ src/switch_rtp.c
@@ -5638,7 +5638,7 @@ static int rtp_common_write(switch_rtp_t *rtp_session,
 	if (send_msg) {
 		bytes = datalen;
 
-		m = send_msg->header.m;
+		m = (uint8_t) send_msg->header.m;
 
 		if (flags && *flags & SFF_RFC2833) {
 			send_msg->header.pt = rtp_session->te;
